\name{cardinal-data}
\docType{data}

\alias{cardinal}
\alias{cardinal.peaklist}
\alias{cardinal.peaks}
\alias{data:cardinal}
\alias{data:cardinal.peaklist}
\alias{data:cardinal.peaks}

\title{Cardinal Painting}

\description{
	DESI-imaging mass spectra collected from the oil painting of a cardinal. Should be loaded with \code{exampleMSIData()}.
}

\source{
    Aston Labs (Livia S. Eberlin, Christina Ferreira, and R. Graham Cooks).
}

\examples{
cardinal <- exampleMSIData("cardinal")
cardinal
}

\keyword{datasets}
