\name{pig206-data}
\docType{data}

\alias{pig206}
\alias{pig206.peaklist}
\alias{pig206.peaks}
\alias{data:pig206}
\alias{data:pig206.peaklist}
\alias{data:pig206.peaks}

\title{Cross-Section of a Whole Pig Fetus}

\description{
	DESI-imaging mass spectra collected from the cross-section of a whole pig fetus. Should be loaded with \code{exampleMSIData()}.
}

\source{
    Aston Labs (Livia S. Eberlin, Christina Ferreira, and R. Graham Cooks).
}

\examples{
pig206 <- exampleMSIData("pig206")
pig206
}

\keyword{datasets}
