\name{CardinalWorkflows-package}
\alias{CardinalWorkflows-package}
\alias{CardinalWorkflows}
\docType{package}

\title{Datasets and workflows for the Cardinal mass spectrometry imaging package}

\description{
Datasets and workflows for Cardinal: DESI and MALDI examples including pig fetus, cardinal and farmhouse painting, and human RCC.
}

\details{
CardinalWorkflows provides datasets and example workflows of mass spectrometry imaging experiments using the Cardinal package for MS imaging analysis.

To view the example workflows, type \code{browseVignettes("CardinalWorkflows")}.
}

\author{
Kyle D. Bemis

Maintainer: Kyle D. Bemis <kbemis@purdue.edu>
}
\keyword{package}
